refactor(logger): ban and message logging
